Use PlistBuddy instead of defaults in scripts
The defaults command will sometimes fail with an error like this:

> The domain/default pair of (.../FraudForce.framework/Info.plist, CFBundleExecutable) does not exist

The defaults command seems intended for use with the macOS preferences system. Included in its man page is:

> WARNING: The defaults command will be changed in an upcoming major release to only operate on preferences domains. General plist manipulation utilities will be folded into a different command-line program.

Perhaps Catalina is that upcoming major release 😅

A better alternative to using defaults is PlistBuddy. If you replace the defaults command in slim-build-frameworks.sh with /usr/libexec/PlistBuddy -c "Print :CFBundleExecutable" "${OUTPUT_FRAMEWORK_PATH}/Info.plist" then it will return the same value and shouldn't fail with the error quoted above.